Correcting the catalog URL (#801)
* Correcting the catalog URL

The URL of where to find dv360 as a destination in the catalog was incorrect. I corrected it and added a line about how to find the values customers are supposed to input.

* Modifying to combine information

Per comment on PR

* Make catalog URL clearer

Correcting my sentence to hopefully minimize confusion.

* Make description of catalog URL clearer

Last correction! Included the formatting to make the description on how to find the slugs clearer

* Update src/connections/destinations/catalog/personas-display-video-360/index.md

Co-authored-by: LRubin <sanscontext@users.noreply.github.com>