Edit SFMC doc
The main change I made this PR for is because we want this destination to already be made in the workspace before they write into us to enable batching. They will need to provide the link to that destination before we can enable anything so I hopefully made that clearer but let me know if this no longer makes sense. Otherwise, just some minor spell check changes and taking out duplicate sentences!