Add userId is required for audiences/computed traits
Received request from Solutions Architect extraordinaire Joe Ayoub:

Hi friends@ team, 

A customer of ours noticed that the Customer.io Destination requires users in a Personas Audience to have userIds. If the users don't have userIds then the user will not be synced to Customer.io (as part of a Personas Audience). 

There is no mention of this requirement in the Personas documentation. 
It essentially means that it is not possible to sync anonymous Audiences to Customer.io, even if all the users have email addresses. This is important information for our customers to have. 

Would it be possible to update the Destination docs to state that a userId is required when syncing an Audience or Computed Trait? Obviously, an email address is important when syncing Audiences to Customer.io (but I don't know if email is actually required...)

Thanks,
Joe