Updates about session-id timestamp
A customer reached out to friends@segment with the following info:

We discovered after implementation that the timestamp value that passes via session_id needs to be unix format; if it is ISO 8601 for instance this generates an error upon delivery to Amplitude (and the event will not be delivered). This requirement is not clearly described in the above documentation so wanted to relay so it can be updated.

--------------

This seemed like a good addition to the docs so other customers could avoid running into this issue :-).