Mixpanel Autotrack Deprecated
Mixpanel reached out to Success Eng to ask us to remove references to Autotrack as it has been disabled. In our doc we have two references, one under features and the other under appendices. I updated the first reference (under features) to advise of this change, and removed the section under Appendices.