Add note about pageName screen setting.md
We added a new setting a few months ago and while it appears under the "Settings" in our docs, it is not explicitly explained. This adds an example of how the setting works. Here is the PR for the original change: https://github.com/segmentio/integrations/pull/1722 and JIRA for release last December: https://segment.atlassian.net/browse/CC-7103.