correcting usage of billing terms
My copy previously indicated that "each object ingested counts as 1 throughput". That's not really accurate as throughput is defined as `(objects ingested + API calls received) / MTU allowance`.  The language is really tortured here, mostly because the definition of throughput is so complex. Not sure how to make this clearer and more readable, open to suggestions!