Clarify instructions on renaming a schema
The previous instructions were unclear and didn't provide detailed instructions on which tabs to select to get to the options that needed modification.
Please confirm my understanding of the impact of making these changes. It appears as if I disable a source, I lose all data from that source until I re-enable it, is that correct?