Update index.md
Updated the date from June 2 > July 8, 2022

Re-ordered a few items in the tables, but mostly just renamed events after we made the client side > server side switch, we consolidated a few events so now "Training" > "Content" and "Content Started" covers courses that could be started from a course collection, path or certificate which is more in line with the way the data is captured in our own internal reporting. 

We lost the ability to track "Exits Course" as this was a navigation item, so that was removed. Also I am unable to get details on E-Commerce, so those were removed as well. 

Minor tweaks to Event Properties table: adding user. in front of id, email, name. We also aren't doing anything with email subjects, I believe that was errantly carried over from the HackMD template.