Remove action source table
Hi team! I saw this table got added and it's incorrect. Customers can use any field or hardcode the action source in the new Facebook Conversions API (Actions) destination. In addition, we do not default to anything. This note said we default to web. I would like to remove the table and have customers rely on the FB Documentation that we link to in the description of action source so that we do not need to keep this table updated on our end. Seems repetitive to keep this in here, and in it's current state I think it's confusing. Thanks!