Adding docs for Segment RETL destination (#3911)
* Adding docs for Segment RETL destination

* First edits

Going to make more edits, but committing these for now

* Update Segment destination docs for public beta

* Update index.md

* Update to the steps for connecting the Segment destination

Updated the steps for connecting and configuring the Segment destination. The goal of this doc update is to prevent customers from accidentally only sending a full sync to the first mapping created. It is therefore recommended to configure and enable all mappings before enabling the Segment destination.

* Update destination ID

* Set to private

Co-authored-by: Sai Gattu <sai.gattu@segment.com>
Co-authored-by: kdaswani <49517136+kdaswani@users.noreply.github.com>
Co-authored-by: James Reynolds <jameshreynolds@gmail.com>
Co-authored-by: markzegarelli <mark.zegarelli@segment.com>