Add details on Replay when connecting Engage Sources
Adding the replay details in this section provides additional context around what occurs when a Source is added. The common misconception is that all of that source's data will be available in the Engage Space, but having this section added would subvert that misconception and enable customers to reach out preemptively with replay requests for historical data beyond 30 days.