add support clarification for simple_segment
As simple_segment is not owned and maintained by Segment, our customers are welcome to use it but should know that our support team won't be able to provide much assistance in debugging implementations that use it. The support team works exclusively with Segment products and doesn't have the capacity to provide in-depth support for tools that exist outside of Segment.