update to reflect source ID > secret ID change (#4182)
Segment used to connect to Amazon Kinesis Firehose via `sourceId`. Recently, an update was made for security reasons to use a Secret ID that is automatically created in the Settings for each Amazon Kinesis Firehose destination in Segment instead. Our documentation doesn't reflect this which leads to confusion for customers. This change updates our documentation to reflect the new way Segment connects to Firehose through Secret ID.