Update index.md
Modify the "Recording userId" section, add in information to indicate userId should not be pass as 'Id' in trait as trait.id is a reserved trait. It does not get sent to MailChimp.

Added in the following sentences to "Recording UserId" section:
"One thing to take note is not to pass 'userId' as 'id' in trait as trait.id is a reserved trait. You can pass 'userId' in a trait name with a corresponding merge field in MailChimp."