Add additional info regarding timestamp fields (#4456)
* Add additional info regarding timestamp fields

Added a note on Iterable's date field in our document on Iterable destination. Iterable has a specific date format that must be used to segment a field by date. Thus, Segment converts ISO strings to UTC before sending it to iterable.

We have had tickets on this issue where a customer is confused as why the timestamp is getting converted to UTC despite sending the local time.

* Apply suggestions from code review

Co-authored-by: stayseesong <83784848+stayseesong@users.noreply.github.com>

---------

Co-authored-by: markzegarelli <mark.zegarelli@segment.com>
Co-authored-by: stayseesong <83784848+stayseesong@users.noreply.github.com>