Update debugger.md
Added information that the source should not be disabled also.

Original sentence:
This automatically generated source cannot be deleted even when the destination is removed, in order for Segment to function as designed. The source will be reused by Segment as needed.

Added "and should not be disabled" and modified to:
This automatically generated source cannot be deleted and should not be disabled even when the destination is removed, in order for Segment to function as designed. The source will be reused by Segment as needed.