Note about array of objects in the builder.md
Engage does support nested traits, however, accessing objects nested in arrays is not currently supported by the Audience builder.

To go a bit more into how arrays work with Engage, when arrays are sent, they get flattened into strings and stored in our backend. For building audiences based on array traits, all the same conditions that would work on strings will work on the array (since the array is now being treated as a string).

Said in other words, when you try to access a trait within an object in an array in the Audience builder, you won't be able to access any of the nested objects. Instead, you can only use string operations like 'contains' and 'does not contain' to look for individual characters or a set of characters in the string.