Added information that SQL trait name should not be the same name as Identifier
Added the following question and answer to indicate that the name of SQL trait should not be the same name as the identifier.

### Why doesn't the identifier updated by SQL trait shows the correct value found in the column?

Ensure that the name given to the SQL trait is not the same name as the identifier or column name from the query. To use SQL trait to update the identifier, the identifier will need to be a column in the query of your SQL trait. The column name in the query of SQL trait should be the one that Identity Resolution use to generate the identifier.