Add Group ID is for mapping of track event to group key $group_id
In the Troubleshooting section, modify the following (original) statement:

"Ensure that the mapping(s) handling your track events have the field for Group ID mapped to a valid value. By default, this field is mapped to the event variable context.groupId."

To the following that mentioned using of group key $group_id:

"For Mixpanel (Actions) destination that uses $group_id as the group key, ensure that the mapping(s) handling your track events have the field for Group ID mapped to a valid value. By default, this field is mapped to the event variable"context.groupId."