Add faq that date string need to be ISO-8601 for time conditions to work
Added the following FAQ:

## Why is the time condition, such as a rolling time window condition of within last N number of days, configured on a trait not working in my audience?

For time conditions to work on a trait, please check that the date string of the trait is in ISO-8601 format. Further information on the date string requirement is mentioned in [document on Spec: Common Fields regarding timestamp](https://segment.com/docs/connections/spec/common/#timestamps).