Update Profile insights.md
Updating to instruct users how to Filter by External ID Type effectively. It's important to highlight that the specific external ID value must be entered in the search bar for the search to be successful. At present, if the value isn't provided, there's no error message displayed in the user interface; it only appears in the developer tools console, which can be confusing. By clarifying this in the documentation, I hope to streamline user experience by clearly indicating the requirement of this specific input.