Make it more clear how to add an endpoint
We were trying to use the webhook destination and were confused when there was no URL setting. The docs should be more explicit that you add endpoints by creating mappings.

Also added a suggestion to use Svix Play for people who don't have an endpoint ready. Have seen this type of suggestion in other webhook docs and its helpful to see the webhooks in action before having to develop an endpoint.