Add info on how to search for alias in Braze to avoid confusion
Add the following information:
---------------------------------------------------------------
### Missing events
When an event is sent under an alias, the event might seems to be missing when the alias cannot be found in Braze. This could be due to incorrect search for the alias in Braze. To search for alias in Braze, the search should be in the format "Alias Label:Alias Name". For example, if "Alias Label" field is set as email and "Alias Name" field is set as email address e.g. "test@email.com", search for the alias in Braze should be "email:test@email.com"