Add info that timestamp shown in audience builder's UI is based on local timezone
Added the following info in Time Comparison section:
-----------------------------------------------------
While in the audience builder, the timezone you see on the UI will be based on your local timezone but it is converted to UTC on the backend. Pleaae take note of this while you are setting time in the audience builder.