Rate limit Klaviyo
If the issue persists even after enabling batching, it will be auto-tuned. We don't have a fixed rate limit for the Klaviyo destination; our system uses an adaptive algorithm to determine the proper rate at which to send events. If there is an increase in Klaviyo destination rate limit, our system will adapt to a rate that is slightly faster than what the downstream service can handle. The occurrence of 429 and other retryable errors actually signals our egress to slow down. The more retryable errors we encounter, the slower we send events; if we achieve more successes, the rate will increase.