Add information on troubleshooting Group properties not attribute Mixpanel group
Added the following
-----------------------------------------------------------------

### Group properties are not attributed to Mixpanel Groups

If the Mixpanel (Actions) destination uses $group_id as the group key, ensure that the mappings handling your `Group` events have the field for **Group ID** mapped to a valid value and the field for **Group Key** is left blank or without value. By default, the field **Group ID** is map to the event variable `groupId` and leaving the field for **Group Key** as blank will use $group_id as the group key.