Update roles.md
New Information added to PII access:

For a user with PII Access and Source Admin/Read-Only permission, it means it there are PII's in the Sources/Debugger, they would be able to view it. However, this doesn't translate to Privacy Portal access.

The Privacy Portal is only accessible by Workspace owners.