Update RETL Query Length Limit
Hi, I am on the RETL team and we noticed that in our public docs we are using an erroneous value for the query length. For the `query` column in our MySQL DB we use MYSQL `TEXT` which has a character limit of 65,535. When a customer/user saves anything longer than that it auto truncates to that limit, which will lead to the query failing.