ID sync strategy for Audience Exit events.
#### Will the Audience Entered and Exited events obey the ID sync strategy? 
Only the "Audience Entered" event will obey the ID sync strategy. However, for the "Audience Exited" event, the ID Sync configuration and space-level ID strategy are not applied. For Audience Exit, all ID combinations are sent in an effort to try and match any identifier that was ever sent to the downstream destination in order to remove it from the external audience.