Unable to view/edit the mapping as expected, and it is throwing a 'Couldn't load page' error.
It might be due to the browser cache or the event property name containing a "/". To resolve this, try clearing the browser cache or accessing the mapping page in an Incognito window. Additionally, check if the property name mapped in the mapping contains a "/". If so, please try renaming the property name without the "/" and map it in the mapping to avoid this error.